Renames hashtable functions to avoid future conflict with pull from master
